# api/src/main/resources/openapi/admin-api.yaml
openapi: 3.0.3
info:
  title: Song Quotes Admin API
  version: 1.0.0
  description: Administrative endpoints for managing quotes.
servers:
  - url: /
tags:
  - name: Admin
    description: Administrative quote management operations.
paths:
  /admin/quote:
    post:
      tags:
        - Admin
      summary: Create a new quote
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/components/schemas/Quote'
      responses:
        '200':
          description: Quote successfully created
          content:
            application/json:
              schema:
                type: string
  /admin/quotes:
    post:
      tags:
        - Admin
      summary: Create multiple quotes
      operationId: createQuotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: './components/schemas.yaml#/components/schemas/Quote'
      responses:
        '200':
          description: Quotes successfully created
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /admin/quote/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
        - Admin
      summary: Delete an existing quote
      operationId: deleteQuote
      responses:
        '204':
          description: Quote successfully deleted
    put:
      tags:
        - Admin
      summary: Replace an existing quote
      operationId: updateQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/components/schemas/Quote'
      responses:
        '204':
          description: Quote successfully updated
        '400':
          description: Quote contains restricted fields
    patch:
      tags:
        - Admin
      summary: Partially update an existing quote
      operationId: patchQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/components/schemas/Quote'
      responses:
        '204':
          description: Quote successfully updated
        '400':
          description: Quote contains restricted fields
  /admin/quotes/reset-hits:
    post:
      tags:
        - Admin
      summary: Reset hits for all quotes
      operationId: resetQuoteHits
      responses:
        '204':
          description: Hits successfully reset
  /admin/quotes/reset-posts:
    post:
      tags:
        - Admin
      summary: Reset posts for all quotes
      operationId: resetQuotePosts
      responses:
        '204':
          description: Posts successfully reset
  /admin/outbox/worker/batch-size:
    post:
      tags:
        - Admin
      summary: Update the outbox worker batch size
      operationId: updateOutboxWorkerBatchSize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/components/schemas/UpdateOutboxBatchSizeRequest'
      responses:
        '204':
          description: Outbox worker batch size successfully updated
  /admin/export:
    get:
      tags:
        - Admin
      summary: Export all quotes as SQL statements
      operationId: exportQuotes
      responses:
        '200':
          description: SQL export of all quotes
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    Quote:
      $ref: './components/schemas.yaml#/components/schemas/Quote'